requirements.txt
asgiref==3.11.1
asttokens==3.0.1
black==26.3.1
click==8.3.1
colorama==0.4.6
coverage==7.13.4
crispy-bootstrap5==2026.3
decorator==5.2.1
dj-database-url==3.1.2
Django==6.0.3
django-crispy-forms==2.6
django-debug-toolbar==6.2.0
django-jazzmin==3.0.3
executing==2.2.1
flake8==7.3.0
gunicorn==25.1.0
iniconfig==2.3.0
ipython==9.11.0
ipython_pygments_lexers==1.1.1
isort==8.0.1
jedi==0.19.2
matplotlib-inline==0.2.1
mccabe==0.7.0
mypy_extensions==1.1.0
packaging==26.0
parso==0.8.6
pathspec==1.0.4
pillow==12.1.1
platformdirs==4.9.4
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pure_eval==0.2.3
pycodestyle==2.14.0
pyflakes==3.4.0
Pygments==2.19.2
pytest==9.0.2
pytest-django==4.12.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytokens==0.4.1
six==1.17.0
sqlparse==0.5.5
stack-data==0.6.3
traitlets==5.14.3
tzdata==2025.3
wcwidth==0.6.0
whitenoise==6.12.0
